Validate received JSON
Checks that received JSON conforms to expected format. I did
investigate using libraries for this, but it turns out our needs are
very simple at the moment so we can roll our own checking; it's very
simple and works well with Typescript as things stand. I've not made
any attempt at a 'generic' solution here.

This commit also attempts some minor tidyups, for example tidying and
validating data *before* passing it to React.

The transform/validate stage is still a bit messy. Things would be
much simpler if we required clients to pass exactly the data required
and in the required format. That way we could simply clean/validate
without the transform stage.