Introduce special styling/layout for liveblog top meta
Note, to re-use text styles, I've continued the pattern of separate
exports in the Text element component. Longer-term the design
system will help with this kind of stuff, but it feels like a clean
enough and pragmatic solution for now to achieve the re-use we want
without coupling components/overriding styles.