Modify data model
Some additional data model changes to support things missing
from the previous PR (as have now tested more thoroughly and also
implemented a first draft of the related Frontend changes).

Specifically, aims to:

* remove stuff from model that is derived (age warning, sharing
URLs)
* include nav stuff

This is a large change, and there will likely be cleanup after
we're confident it works on production.