Introduce startup wrapper
The intent here is twofold:

1. improve DI

While we have to share behaviour via the window object, module code
itself can be ignorant of this fact is we use regular DI from init
downwards. This should make testing a lot easier and keep things
manageable over time re globals.

2. standardise inits

This improves DX and also allows common behaviour to be applied.
As an initial motivation, performance logging is added in this
commit.