Add a skeleton commercial intro doc
This doc is meant to help the Dotcom team gain a basic
understanding of the Commercial JS on the site during the
migration. We also want to document some basic trouble-
shooting guidance to assist with local development and
production monitoring.

It is recognised that it is likely to go out of date quickly!