Send Ophan VIEW event when epic >50% visible
Also record the view in localstorage at the same time (we were already
doing this, but on render).

The useHasBeenSeen hook is based on this blog post:
https://medium.com/the-non-traditional-developer/how-to-use-an-intersectionobserver-in-a-react-hook-9fb061ac6cb5

The hasBeenSeen logic is wrapped in a useEffect. Although hasBeenSeen
will only change from false -> true once, if anything else triggers a
re-render (including props/state which are introduced in the future)
then without the useEffect we'd end up sending the VIEW event again. I
don’t want to rely on second-guessing when React will render, so this
provides some insurance.