Convert test name to Camel Case
The config.switches object converts from kebeb case to camelCase. The
test config needs updating to reflect this.

In addition, the `CAPI.config.switches` used by `document.tsx` was the
incorrect one for this use case. Instead I've modified the code to take
the active experiment CSS string as an argument into the document
component and to handle the generation and extraction in `render.tsx`