Add guardian.functions.import as an all-browser dynamic import
Dynamic import is useful for loading remote ES modules, but isn't
supported in IE 11 and also older versions of the main browsers.

At the moment ~90% of browsers support it out of the box.

This PR polyfills dynamic import for non-supporting browsers. To
do this it adds a < 2kb (gz) high priority script to all requests.
In addition, for browsers that don't support ES modules (IE 11), a
larger polyfill is dynamically loaded (3kb gzipped).

As browser support improves we can simplify this code.

It's not possible to directly polyfill `import` as it is a keyword
in Javascript.