Hide NoJS input checkbox from screen readers
This patch fixes the google lighthouse audit

    [role]s are not contained by their required parent element
       <input type="checkbox" class="css-92oqeo"
        id="top-nav-input-checkbox" name="more" tabindex="-1"
        role="menuitemcheckbox" aria-checked="false">

Which highlights that an element with the role `menuitemcheckbox`
must have a parent with the `menu` role. In this case, the parent has
the `nav` role. One way to fix this would be switch `nav` to `menu`,
but `nav` is a landmark role that I think we should keep as it's
semanticaly appropriate.

The solution in this patch is to actually hide the checkbox. The
reason is that the checkbox isn't really a checkbox. It's not meant
to be seen or clicked on. It's part of a pattern that makes it
possible to build interactive menus without using javascript. It's
sometimes called the "checkbox hack"
https://css-tricks.com/the-checkbox-hack/

Essentialy, the purpose of the checkbox is to keep state - if the
menu has been toggled on or off. This state change can be used to
dynamicaly change the CSS of the menu.

So this patch hides it from screenreaders, the same it's hidden from
visual and keyboard navigation.

It does this for the nav, but also on the expanded menu itself. Even
though it's not flagged as a problem in lighthouse, this nojs pattern
makes every entry in the expanded menu be read twice by VoiceOVer.