Ensure that we always have a brazeUuid promise
This updates the behaviour around asyncBrazeUuid to ensure that it
always gets set regardless of whether a user is logged in (in this way
it now behaves like asyncCountryCode). The asyncBrazeUuid is a
dependency for the banner picking logic which will now run once this
effect has run (after the first render cycle of App). Previously there
was an issue where asyncBrazeUuid was sometimes not set which prevented
the banner picker from running.