Adds Ophan click tracking for Braze Banner
The Ophan call itself lives in @guardian/braze-components, the platform
supports this by providing submitComponentEvent as a prop to the
component.

I bumped @guardian/types to 0.4.5. This is to fix an issue where
braze-components can't find the new ophan.ts. Even though it has a dev
dependency of types@0.4.5 it's seeing the 0.4.4 version from DCR.