Changes to support multiple Braze banners
Unsubscribe on first in-app message. We only ever want to deliver a
single banner on screen at a time. Any further messages will be ignored
anyway because we’re wrapping this all up in a promise. Unsubscribing
means we get the lower priority message on a subsequent page view.

Update the pre check to look at gu_hide_support_messaging cookie. This
widens the base of users who we'll load the Braze SDK for since the new
canvas audience is larger. This works because users who have the
gu_hide_support_messaging cookie have a product or have otherwise
contributed.