Refactor dependency handling for the Braze SDK
Refactors the way dependencies for loading the Braze web SDK are
handled. By dependencies we mean the information we need and the
conditions which need to be true before we can load the SDK.

Previously the dependency checking was intermingled with other code in
BrazeBanner. We think this new approach is much clearer and it's easier
to write tests around. We'd like to extract more Braze related code out
to its own NPM package, which could be shared across DCR and frontend.
Modularising the existing code is a first step along that path. It's
also much easier to have functionality like common logging for when
dependencies aren't all met which is really useful when developing.