Replacing explicit height and width with role on ClickToView component
Ideally the click to view component would be exactly the same size as the
embedded content. It turns out its not possible to calculate this size
without actually loading the embedded content which negates the
reason for having the overlay in the first place.

So the overlay will render to its natural size and will expand/shrink
to reveal the embedded content when clicked.

This is not ideal for ux purposes but should not affect page score:

See:https://web.dev/cls/#user-initiated-layout-shifts