Replacing calls to embedIframe with calls to updateIframeHeight in ClickToView.stories
embedIframe rewrites the srcDoc for all the iframes created by the UnsafeEmbedBlockComponent
this causes all of those iframes to be refreshed when each ClickToView wrapped is clicked on,
which isnt ideal.

Using updateIframeHeight with a selector for the specific iframe being clicked on
ensures that just that iframe is effected.