Use an 'immersive' card style for Carousel test
This adds a `isFullCardImage` prop that is used to create an 'immersive'
card component for use in the Carousel AB test.

This will definitely need refactoring  at a future date to make this
much easier to read and handle, but for an AB test this _should_
consolidate all of the design changes required into a single commit that
can be reverted if not needed. If this gets taken into production, we
will probably want to update this as part of the work to rationalise
card designs.