Add image sources for multiple card sizes
The Carousel has an immersive design variant which requires different
image sizes based on the size of the Card. Rather than passing one big
image that than gets compressed down, we get a set of images at
different sizes in the trail. As long as these are available and the
carousel should be immersive, we use a `picture` element to hanbdle the
different image sizes, with the final fallback being the thumbnail.