Remove picture element
Remove the `<picture>` element from `<Card>` so that we only implement
this when we're ready to do it fully and not partially as was done here.

Instead, use `<img>` but for large carousels only pass through the
`460px` image. This isn't ideal long term but will work for the AB test.
This means that we can revisit this post-test and implement this
properly.