Integration of CAPIFormat into DCR
Initial integration of the new CAPIFormat type into DCR. This primarily
fixes the use of the different `decide...` functions to use the format
data passed through from CAPI. There are also updates to fixtures and
stories done so as to pass typescript compilation.

The previous `pillar` and `designType` fields are kept on the `CAPIType`
and in a few other places to maintain parity with the JSON data passed
from Frontend. Once this integration is in place, those fields can be
removed in the sent data and then removed here as well. At that point
there will be a second round of integration, focussed primarily on
pruning of old information.