Update cypress and fixtures
The Cypress version and a number of fixtures were out of date. In
particular, fixture data for most-read and most-read-geo were both out
of date by a number of months and were missing a number of important
backend data changes.

To fix this I've
- Bumped Cypress to the latest minor version.
- Updated the Rich Link JSON file
- Moved to using the Jest/Storybook fixture data for most read and most
read geo by improting it directly into `mocks.js`